Moved populate_neo4j to its own file
