+++ tests/e2e/pinecone_e2e/populate_dbs.py
@@ -12,12 +12,15 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+
 import hashlib
 import json
 import os.path
 
 from neo4j import GraphDatabase
 from pinecone import Pinecone, ServerlessSpec
+
+from ..populate_neo4j import populate_neo4j
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 
@@ -37,39 +40,6 @@ def populate_dbs(neo4j_driver, pc_client, index_name="jeopardy"):
 
     # Populate Neo4j
     populate_neo4j(neo4j_driver, neo4j_objects)
-
-
-def populate_neo4j(neo4j_driver, neo4j_objs):
-    question_nodes = list(
-        filter(lambda x: x["label"] == "Question", neo4j_objs["nodes"])
-    )
-    answer_nodes = list(filter(lambda x: x["label"] == "Answer", neo4j_objs["nodes"]))
-    category_nodes = list(
-        filter(lambda x: x["label"] == "Category", neo4j_objs["nodes"])
-    )
-    belongs_to_relationships = list(
-        filter(lambda x: x["type"] == "BELONGS_TO", neo4j_objs["relationships"])
-    )
-    has_answer_relationships = list(
-        filter(lambda x: x["type"] == "HAS_ANSWER", neo4j_objs["relationships"])
-    )
-    question_nodes_cypher = "UNWIND $nodes as node MERGE (n:Question {id: node.properties.id}) ON CREATE SET n = node.properties"
-    answer_nodes_cypher = "UNWIND $nodes as node MERGE (n:Answer {id: node.properties.id}) ON CREATE SET n = node.properties"
-    category_nodes_cypher = (
-        "UNWIND $nodes as node MERGE (n:Category {id: node.id}) ON CREATE SET n = node"
-    )
-    belongs_to_relationships_cypher = "UNWIND $relationships as rel MATCH (q:Question {id: rel.start_node_id}), (c:Category {id: rel.end_node_id}) MERGE (q)-[r:BELONGS_TO]->(c)"
-    has_answer_relationships_cypher = "UNWIND $relationships as rel MATCH (q:Question {id: rel.start_node_id}), (a:Answer {id: rel.end_node_id}) MERGE (q)-[r:HAS_ANSWER]->(a)"
-    neo4j_driver.execute_query(question_nodes_cypher, {"nodes": question_nodes})
-    neo4j_driver.execute_query(answer_nodes_cypher, {"nodes": answer_nodes})
-    neo4j_driver.execute_query(category_nodes_cypher, {"nodes": category_nodes})
-    neo4j_driver.execute_query(
-        belongs_to_relationships_cypher, {"relationships": belongs_to_relationships}
-    )
-    res = neo4j_driver.execute_query(
-        has_answer_relationships_cypher, {"relationships": has_answer_relationships}
-    )
-    return res
 
 
 def populate_pinecone(pc_client, pc_question_objs, index_name):
