+++ tests/e2e/pinecone_e2e/test_pinecone_e2e.py
@@ -22,8 +22,9 @@ from neo4j_genai.retrievers.external.pinecone import PineconeNeo4jRetriever
 from neo4j_genai.types import RetrieverResult, RetrieverResultItem
 from pinecone import Pinecone
 
+from ..populate_neo4j import populate_neo4j
 from ..utils import EMBEDDING_BIOLOGY
-from .populate_dbs import build_data_objects, populate_neo4j
+from .populate_dbs import build_data_objects
 
 
 @pytest.fixture(scope="module")
