+++ tests/e2e/weaviate_e2e/populate_dbs.py
@@ -13,14 +13,17 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
+import hashlib
+import json
+import os.path
+from typing import Any
+
 import weaviate.classes as wvc
+from neo4j import Driver, GraphDatabase
 from weaviate.client import Client
 from weaviate.connect.helpers import connect_to_local
-from typing import Any
-from neo4j import GraphDatabase, Driver, EagerResult
-import os.path
-import json
-import hashlib
+
+from ..populate_neo4j import populate_neo4j
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 
@@ -47,46 +50,7 @@ def populate_dbs(
     populate_neo4j(neo4j_driver, neo4j_objects)
 
 
-def populate_neo4j(
-    neo4j_driver: Driver, neo4j_objs: dict[str, list[wvc.data.DataObject]]
-) -> EagerResult:
-    question_nodes = list(
-        filter(lambda x: x["label"] == "Question", neo4j_objs["nodes"])
-    )
-    answer_nodes = list(filter(lambda x: x["label"] == "Answer", neo4j_objs["nodes"]))
-    category_nodes = list(
-        filter(lambda x: x["label"] == "Category", neo4j_objs["nodes"])
-    )
-    belongs_to_relationships = list(
-        filter(lambda x: x["type"] == "BELONGS_TO", neo4j_objs["relationships"])
-    )
-    has_answer_relationships = list(
-        filter(lambda x: x["type"] == "HAS_ANSWER", neo4j_objs["relationships"])
-    )
-    question_nodes_cypher = "UNWIND $nodes as node MERGE (n:Question {id: node.properties.id}) ON CREATE SET n = node.properties"
-    answer_nodes_cypher = "UNWIND $nodes as node MERGE (n:Answer {id: node.properties.id}) ON CREATE SET n = node.properties"
-    category_nodes_cypher = (
-        "UNWIND $nodes as node MERGE (n:Category {id: node.id}) ON CREATE SET n = node"
-    )
-    belongs_to_relationships_cypher = "UNWIND $relationships as rel MATCH (q:Question {id: rel.start_node_id}), (c:Category {id: rel.end_node_id}) MERGE (q)-[r:BELONGS_TO]->(c)"
-    has_answer_relationships_cypher = "UNWIND $relationships as rel MATCH (q:Question {id: rel.start_node_id}), (a:Answer {id: rel.end_node_id}) MERGE (q)-[r:HAS_ANSWER]->(a)"
-    neo4j_driver.execute_query(question_nodes_cypher, {"nodes": question_nodes})
-    neo4j_driver.execute_query(answer_nodes_cypher, {"nodes": answer_nodes})
-    neo4j_driver.execute_query(category_nodes_cypher, {"nodes": category_nodes})
-    neo4j_driver.execute_query(
-        belongs_to_relationships_cypher, {"relationships": belongs_to_relationships}
-    )
-    res = neo4j_driver.execute_query(
-        has_answer_relationships_cypher, {"relationships": has_answer_relationships}
-    )
-    return res
-
-
-def populate_weaviate(
-    w_client: Client,
-    w_question_objs: list[wvc.data.DataObject],
-    collection_name: str,
-) -> None:
+def populate_weaviate(w_client, w_question_objs, collection_name):
     questions = w_client.collections.get(collection_name)
     questions.data.insert_many(w_question_objs)
